Update nuget.json
I know the nuget.exe is can update itself, but the exe that's included in the nuget.org site is actually not up to date, so when you install the nuget package you have an old version.  This at least seems like a better way of keeping track of nuget versions.