FIX `bash` startup failure (within GOW)
* create the `etc` directory (if needed)

.# DISCUSSION

`bash` requires the `etc` directory for startup. However, `etc` is missing
from the GOW repo. There are current pull requests fixing this issue (see
bmatzelle/gow#192). But, this fixes the the issue while waiting for the GOW
repo to be updated. The fix here is compatible with all of the current possible
future GOW repo fixes.