Fixed an issue with the 32/64 bit versions of R
The R installer delivers the 32 as well as the 64 bit versions of R. I
added a pre_install to r.json to choose the correct version depending on
the user's architecture.

Additionally I added a new shim for 'rterm.exe' to have an alternative
way to start R without removing the 'r' alias in powershell.