Refactor perl manifest to use bin
Currently, every executable under the moon is available via the PATH when installing the perl manifest. This creates issues when compiling C/C++ code since build tools find the compilers/linkers included in perl manifest and decide to use those over other installed compilers.

I changed the manifest to only add perl.exe and wperl.exe to the PATH. It is extremely likely that there are other useful executables buried in the perl bin directory but it seems like a reasonable approach to add these on a as-needed basis when issues are reported by users who expect some binary to be available from the perl manifest (compared to polluting the PATH of everyone who installs the perl manifest only to have perl.exe available).

This is a breaking change but I feel the current situation is rather excessive and should be fixed.