remove dependency on git installed through scoop
currently, if I don't have `git` or `git-with-openssh` installed through scoop I get a confusing error indicating `git` couldn't be found, even though it is right there on my path

there are many apps that come bundled with git now, including github desktop, Cmder etc. it would be better for check for these on path as a last resort, and work out the location to `bash.exe` from there.