Fix installation path for executable files
In Version 1.12.0 the binary assets for hyperfine are not in the top level directory of the zip release file - in contrast to what was the case the former versions.
Therefore, regular install does not find the executable and thus cannot create the shim. 
Proposed change mitigates the issue by adjusting extraction path.