azure-cli: Fix AzureCLIPath environment variable (#2717)
The 2 AzureCLIPath environment variables introduced in https://github.com/ScoopInstaller/Main/pull/1469 are not set to the actual az cli path. Since az is installed in the \wbin sub folder, as seen in the bin property. 

This for example resulted in tools using the go-autorest library not working when the az cli is installed using scoop. 
https://github.com/Azure/go-autorest/blob/c7f947c0610de1bc279f76e6d453353f95cd1bfa/autorest/azure/cli/token.go#L128