ImageMagick - includes DLLs for programming APIs
The manifest now downloads the full installer and unpacks it rather than the portable version. It also adds the environment variables and the installation folder to the path. This allows other programs such as Python Wand to find the programs and libraries for their own use. If further environment variables are missing, they can be added from this list: https://imagemagick.org/script/resources.php#environment. The bundled ffmpeg was unused in the portable version but as the manifest now adds the installation directory to the path, the outdated ffmpeg is renamed so that users can continue using the up-to-date ffmpeg downloaded by scoop.