ImageMagick: includes DLLs for programming APIs (#2885)
* ImageMagick - includes DLLs for programming APIs

The manifest now downloads the full installer and unpacks it rather than the portable version. It also adds the environment variables and the installation folder to the path. This allows other programs such as Python Wand to find the programs and libraries for their own use. If further environment variables are missing, they can be added from this list: https://imagemagick.org/script/resources.php#environment. The bundled ffmpeg was unused in the portable version but as the manifest now adds the installation directory to the path, the outdated ffmpeg is renamed so that users can continue using the up-to-date ffmpeg downloaded by scoop.

* Change ffmpeg dependency to suggest, add note for info

* Remove convert.exe to stop conflict with built-in command

* Remove shims because already in path, account for convert.exe conflict

* Removed suggest ffmpeg and improved notes

* Restored mirror urls

* Keep shims just in case, add note that convert is deprecated

* Update notes

Co-authored-by: Rashil Gandhi <46838874+rashil2000@users.noreply.github.com>