Escape single quotes and spaces in the path of neofetch fed as an argument to bash.exe in "pre_install" property
Many windows usernames contain spaces or single quotes/apostrophe. I accidentally have an apostrophe (') in my username. As a result neofetch fails to run as mentioned in detail [here](https://github.com/dylanaraps/neofetch/issues/2002#issue-1084846469)

I discovered that the fix to this issue is to surround the path of the neofetch script being passed as an argument to bash.exe in double-quotes. This would escape spaces and single quotes in username. Windows does not allow double quotes or backslashes in usernames, so there is no need to consider those.

This is the updated working script: https://pastebin.com/J0zTeNAw
I tried to escape my double quotes in the json file with backslashes, please check if it corresponds to my intent.