bitwarden-cli: Update repo and version (#3797)
Bitwarden consolidated all of their clients (except the mobile application) into a single repository, https://github.com/bitwarden/clients.  All current and future releases of bitwarden-cli will be released there.  As we can no longer assume that a new release on that repo is for the CLI, checkver logic was updated to parse the releases.atom feed. This updated manifest allows scoop to offer current versions of the Bitwarden CLI.