sftpgo.json: copy native binary to root
Currently the app comes with 3 architectures: arm64, 32bit and 64bit. However arm64 and 32bit binaries exists in their respective folders which requires users of such architectures to explicitly specify the directory.

With this commit, if needed, the default 64bit binary is backed up to "x64" folder and the native binary for the platform is copied to the root directory, effectively replacing the binary the shim is created for.