sftpgo: Persist configuration across updates (#4135)
* sftpgo: persist configuration across updates

The following files have to be copied to the new version's root directory every time the app gets updated:

id_ecdsa
id_ecdsa.pub
id_ed25519
id_ed25519.pub
id_rsa
id_rsa.pub
sftpgo.db
sftpgo.json

By adding the files above to persist, the config files get copied automatically.

* fix: do not persist ssh keys

Since Windows 10 builds earlier than 1809 do not have a built-in SSH client, the keys cannot be created in a way that doesn't require the installation of an SSH client at pre_install. So I'm only adding the config files to persist and none of the keys.

* sftpgo.json: copy native binary to root

Currently the app comes with 3 architectures: arm64, 32bit and 64bit. However arm64 and 32bit binaries exists in their respective folders which requires users of such architectures to explicitly specify the directory.

With this commit, if needed, the default 64bit binary is backed up to "x64" folder and the native binary for the platform is copied to the root directory, effectively replacing the binary the shim is created for.

* Update sftpgo.json

Co-authored-by: Rashil Gandhi <46838874+rashil2000@users.noreply.github.com>