hledger@1.32.2: Shim `hledger-ui.exe` (#5402)
* update hledger: add hledger-ui shim

hledger-ui is a tui for hledger added for the Windows platform in the latest release 1.32.2.  Unfortunately, the hledger-ui binary was initially overlooked in the .zip file.  This has been fixed although the version number has was not updated.

* hledger: add hledger-ui.exe shim

The initial release of hledger 1.32.2 did not include heldger-ui.exe in the zip.