zigup: (chore) Add architecture, persist zig versions
Add architecture because zigup is 64bit.
Persist downloaded versions of zig which are kept in `$dir\zig`.
Add $dir to PATH because the symlink to `zig` is created here.
Add note that the `zig` symlink needs to be recreated after update. We should use `--path-link` to move this to a folder that Scoop can persist.