mill: Fix urls and update to 1.0.4
The github releases are not longer including the wrapper/install .bat
script. I used the maven repository link referenced in some of their
manual installation method documentation.
Also updated the homepage url to a working one.

Fairly important that this works, as their documentation references
scoop as an installation method.

Relates to #773