iroh: deprecate manifest (#7120)
This manifest has been error-ing during Excavator action for who knows
how long.

The version in the scoop bucket is very out of date, and looking at
their website documentation at https://www.iroh.computer/ , it seems to
me that they have pivoted to being more of a rust library than a
definitive service that you run as a CLI application. They do not seem
to advertise a app or service to download and install and execute.